[Refactor]: CustomOAuth2UserService 리팩토링
패키지 위치 변경
유틸 클래스 활용하도록 변경
기본 이미지 생성 로직을 이미지 서비스에서 담당하도록 변경

[#?]
